Fixing and reorganizing tests
