+++ c3po-core/src/test/kotlin/core/model/AppStateManagerTest.kt
@@ -1,7 +1,5 @@
-/* TODO: Fix compilation errors
-import core.model.Action
-import core.model.AppState
-import core.model.AppStateManager
+package core.model
+
 import dev.amaro.sonic.IAction
 import dev.amaro.sonic.IReducer
 import io.mockk.every
@@ -62,4 +60,3 @@ class AppStateManagerTest {
         verify(exactly = 1) { stateManager.perform(scheduledAction) }
     }
 }
-*/
