+++ c3po-desktop/src/test/kotlin/SignatureResponseParserTest.kt
@@ -1,11 +1,10 @@
-/* TODO: Fix compilation errors
 import assertk.all
 import assertk.assertThat
 import assertk.assertions.isEqualTo
 import assertk.assertions.prop
-import facade.SignatureResponseParser
-import models.SignatureInfo
-import models.Signer
+import core.facade.SignatureResponseParser
+import core.model.SignatureInfo
+import core.model.Signer
 import org.junit.jupiter.api.Test
 
 class SignatureResponseParserTest {
@@ -36,4 +35,4 @@ class SignatureResponseParserTest {
         }
     }
 }
-*/
+
